# Optimize unicode_normalize methods