… lib/unicode_normalize/tables.rb