This PR makes several changes to the implementation of the internal `UnicodeNormalize` module to improve performance and reduce memory allocations.

- this PR adds an early return for ASCII-only strings – `String#ascii_only?` is extremely fast and helps in cases when user tries to normalize ASCII-only strings (for example, this helps to save a lot of time and memory in the [IDNA mapping from UTS46](https://www.unicode.org/reports/tr46/#ProcessingStepNormalize) which I want to propose as a default IDNA algorithm to [ruby/uri](https://github.com/ruby/uri/issues/76), but that's a different story 😅)
- biggest change in terms of algorithm is that now we use codepoints instead of chars, this helps save on memory allocations (these changes affect only cold cache)
- finally, it makes some small adjustments to reduce memory allocation (`map` + `join` -> `each`, `gsub` -> `gsub!`)

### Benchmark

Benchmark is kinda spoiled by the usage of cache, but still 😅 

Macbook 13", M1, 2020, 16 GB, Ruby 3.2.2

```ruby
require 'benchmark/ips'

value = "google.com"
uni = "\u0136\uAC01\u1100\u1161\u11A8\u1100\u1176\u11a8\uae30\u11a7\uae30\u11c3\uAC00\u0323\u0300\uAC00\u0323\u0300\u1100\u1161\u0323\u0300\u1100\u1161\u0323\u0300\u4EE4\u548C"

Benchmark.ips do |x|
  x.report("orig normalize ascii") {
    value.unicode_normalize
  }
  x.report("orig normalize unicode") {
    uni.unicode_normalize
  }
end
Object.send(:remove_const, 'UnicodeNormalize')

require_relative '../ruby/lib/unicode_normalize/normalize'

Benchmark.ips do |x|
  x.report("new normalize ascii") {
    UnicodeNormalize.normalize(value)
  }
  x.report("new normalize unicode") {
    UnicodeNormalize.normalize(uni)
  }
end
```

```
Warming up --------------------------------------
orig normalize ascii   100.632k i/100ms
orig normalize unicode
                        20.641k i/100ms
Calculating -------------------------------------
orig normalize ascii    981.274k (± 6.2%) i/s -      4.931M in   5.053755s
orig normalize unicode
                        226.720k (± 1.9%) i/s -      1.135M in   5.009079s
Warming up --------------------------------------
 new normalize ascii   802.191k i/100ms
new normalize unicode
                        22.688k i/100ms
Calculating -------------------------------------
 new normalize ascii      8.106M (± 0.9%) i/s -     40.912M in   5.047787s
new normalize unicode
                        226.993k (± 1.2%) i/s -      1.157M in   5.098191s
```

### Memory profiling:

```ruby
require 'memory_profiler'

value = "google.com"
uni = "\u0136\uAC01\u1100\u1161\u11A8\u1100\u1176\u11a8\uae30\u11a7\uae30\u11c3\uAC00\u0323\u0300\uAC00\u0323\u0300\u1100\u1161\u0323\u0300\u1100\u1161\u0323\u0300\u4EE4\u548C"
N = 100_000

puts "Before ASCII-only"
MemoryProfiler.report do
  N.times {
    value.unicode_normalize
  }
end.pretty_print(detailed_report: false)

puts "Before Unicode"
MemoryProfiler.report do
  N.times {
    uni.unicode_normalize
  }
end.pretty_print(detailed_report: false)

Object.send(:remove_const, 'UnicodeNormalize')
require_relative '../ruby/lib/unicode_normalize/normalize'

puts "After ASCII-only"
MemoryProfiler.report do
  N.times {
    UnicodeNormalize.normalize(value)
  }
end.pretty_print(detailed_report: false)

puts "After Unicode"
MemoryProfiler.report do
  N.times {
    UnicodeNormalize.normalize(uni)
  }
end.pretty_print(detailed_report: false)
```

```
Before ASCII-only
Total allocated: 6165812 bytes (136505 objects)
Total retained:  750726 bytes (10771 objects)

Before Unicode
Total allocated: 306554744 bytes (2400094 objects)
Total retained:  200 bytes (5 objects)

After ASCII-only
Total allocated: 0 bytes (0 objects)
Total retained:  0 bytes (0 objects)

After Unicode
Total allocated: 306401688 bytes (2400027 objects)
Total retained:  120 bytes (3 objects)
```