@duerst 2023-11-28
For benchmarking, I'd be very interested in the results of using test/test_unicode_normalize.rb. It runs all the tests provided in the NormalizationTest.txt file (see https://unicode.org/Public/UCD/latest/ucd/NormalizationTest.txt). I'm not sure how difficult it is to extract the actual processing part (not the part reading in the data) for a benchmark.