@skryukov 2023-11-28
@duerst, sure!

[Here is a simple benchmark](https://gist.github.com/skryukov/e1593b8d5cbec4b09d2eeed611e84ed7). I simply ran through every normalization form for every case from the `NormalizationTest.txt`.


Results running on Macbook 13", M1, 2020, 16 GB, Ruby 3.2.2:
```
Rehearsal -----------------------------------------------------
native #normalize   4.193533   0.032114   4.225647 (  4.233881)
new #normalize      4.300044   0.031065   4.331109 (  4.491655)
-------------------------------------------- total: 8.556756sec

                        user     system      total        real
native #normalize   4.366779   0.033569   4.400348 (  4.482564)
new #normalize      4.122924   0.024987   4.147911 (  4.153229)
```

```
Calculating -------------------------------------
   native #normalize     1.421B memsize (   722.440k retained)
                        20.221M objects (    18.001k retained)
                        50.000  strings (    50.000  retained)
      new #normalize     1.231B memsize (   738.520k retained)
                        16.003M objects (    18.463k retained)
                        50.000  strings (    50.000  retained)
```

```
Rehearsal -------------------------------------------------------
native #normalized?   4.736180   1.333103   6.069283 (  6.614149)
new #normalized?      3.521938   1.143961   4.665899 (  4.860499)
--------------------------------------------- total: 10.735182sec

                          user     system      total        real
native #normalized?   3.710057   0.140508   3.850565 (  3.856705)
new #normalized?      3.121058   0.051271   3.172329 (  3.177075)
```

```
Calculating -------------------------------------
   native #normalize   760.305M memsize (   740.920k retained)
                        10.954M objects (    18.463k retained)
                        50.000  strings (    50.000  retained)
      new #normalize   616.285M memsize (   720.040k retained)
                         7.883M objects (    18.001k retained)
                        50.000  strings (    50.000  retained)
```